move color into math and delete colorutil
