--- src/client/scripts/esm/chess/util/colorutil.ts
@@ -1,17 +1,0 @@
-
-/**
- * This scripts contains utility methods for working with players,
- * such as getting the opposite color, trimming the color from type, etc.
- * 
- * ZERO depancies.
- */
-
-
-// Type Definitions -----------------------------------------------------------------------------
-
-/** A color in a length-4 array: `[r,g,b,a]` */
-type Color = [number,number,number,number];
-
-export type {
-	Color
-};
