+++ src/client/scripts/esm/components/header/pieceThemes.ts
@@ -1,7 +1,7 @@
 import { rawTypes, players } from "../../chess/util/typeutil.js";
 
 import type { RawType, Player } from "../../chess/util/typeutil.js";
-import type { Color } from "../../chess/util/colorutil.js";
+import type { Color } from "../../util/math.js";
 
 type PieceColorGroup = {
 	// eslint-disable-next-line no-unused-vars
