+++ src/client/scripts/esm/components/header/themes.ts
@@ -4,7 +4,7 @@
 import jsutil from "../../util/jsutil.js";
 import { players } from "../../chess/util/typeutil.js";
 
-import type { Color } from "../../chess/util/colorutil.js";
+import type { Color } from "../../util/math.js";
 import type { PieceColorGroup } from "./pieceThemes.js";
 
 /*
